requirements.txt
langchain>=1.0.2
# langchain-openai>=1.0.1
langgraph>=1.0.1
pandas==2.2.2
pytest>=8.4.2
python-dotenv>=1.2.0
pyyaml>=6.0.3
ruff>=0.14.2
langchain-google-genai==3.0.0